third_party/WebKit/LayoutTests/http/tests/security/referrer-policy-attribute-import-no-referrer-expected.txt

CONSOLE MESSAGE: line 6: PASS
